Split test_plugins.py into a directory mirroring the source layout
The plugin test file mirrored the monolithic _plugins.py and had grown
to ~360 lines covering five plugins. Now each plugin's tests live in
their own file under tests/test_plugins/, matching dead_cst/_plugins/.

The two helpers (_write, _reachable_fqnames) become fixtures
(write_files, reachable_fqnames) in a local conftest.py so the new
files share them via dependency injection rather than imports. An empty
__init__.py turns the directory into a package, avoiding pytest
basename collisions if a future top-level test file picks the same
name.
